Checking to make sure artefact/block library still exist (Bug #1319634)
Sometimes views will have artefacts/block instances on them but the
plugin has been deleted. This should not stop a user being deleted.

To test:
Add a plugin to the site (eg openbadgesdisplayer block plugin, or cpds
artefact plugin) and install it.
Add a page and add a block to the page.
Delete the dir from the site for the plugin
Try to bulk delete the user

+++ htdocs/lib/user.php
@@ -1364,8 +1364,10 @@ function delete_user($userid) {
     if ($artefactids) {
         foreach ($artefactids as $artefactid) {
             try {
-                $a = artefact_instance_from_id($artefactid);
-                $a->delete();
+                $a = artefact_instance_from_id($artefactid, true);
+                if ($a) {
+                    $a->delete();
+                }
             }
             catch (ArtefactNotFoundException $e) {
                 // Awesome, it's already gone.
